For all caps enum, generating all caps enum
Previously, we lowercase the first letter for enums, so for things like
NewHope, our enum will be newHope, which matches what Swift reference.

But in GraphQL, they prefer all caps, thus, NEWHOPE will be nEWHOPE, and
this is not reasonable.

This PR skips the lowercase logic if it is all cap.