Balance waits in SQLiteConnectionPool drain (#2)
We are encountering a crash when SQLiteConnectionPool is deinitialized. It seems that we need to balance waits with signals in drain as stated in https://developer.apple.com/documentation/dispatch/dispatchsemaphore/1452955-init
```
Calls to signal() must be balanced with calls to wait(). Attempting to dispose of a semaphore with a count lower than value causes an EXC_BAD_INSTRUCTION exception.
```