Add skeleton for persisted, thread-safe dictionary
This commit added the skeleton for a persisted, thread-safe dictionary.
This is useful when we want to store simple key-value objects into
Workspace without creating schema for every single of them. There are
some prerequisite work, in particular, I want to be able to encode not
only Codable, but also FlatBuffers objects into the key-value store.
That requires me to clean up the codegen a bit to allow
FlatBuffersCodable protocol.

This WorkspaceDictionary when it is done, should be a drop-in
replacement for UserDefaults.