Explicitly close connection & move to rollback journal when truncating
For reader pool, the connection will be implicitly closed. However, it
is better to explicitly close it otherwise when there are other places
holding Workspace (even if it is shutdown), we cannot rename / move the
file around.

Also, we cannot move file around without -wal file if it is in wal mode
even if -wal file is 0 in length. Moving to rollback journal when
truncating solves this issue (and it will be wal again anyway when we
open it next time).